feat(vercel): deprecate entrypoints
