+++ packages/vercel/test/fixtures/serverless-with-dynamic-routes/package.json
@@ -4,6 +4,6 @@
   "private": true,
   "dependencies": {
     "@astrojs/vercel": "workspace:*",
-    "astro": "^5.0.0-alpha.8"
+    "astro": "^5.0.0-beta.5"
   }
 }
