+++ packages/vercel/test/prerendered-error-pages.test.js
@@ -15,7 +15,7 @@ describe('prerendered error pages routing', () => {
 
 	it('falls back to 404.html', async () => {
 		const deploymentConfig = JSON.parse(await fixture.readFile('../.vercel/output/config.json'));
-		assert.deepEqual(deploymentConfig.routes.at(-1), {
+		assert.deepEqual(deploymentConfig.routes.find(r => r.status === 404), {
 			src: '/.*',
 			dest: '/404.html',
 			status: 404,
